backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
gunicorn==23.0.0
google-generativeai==0.8.5
